+++ org/scec/sha/gui/beans/IMR_GuiBean.java
@@ -274,8 +274,8 @@ public class IMR_GuiBean extends ParameterListEditor
      StringBuffer b = new StringBuffer();
 
      try{
-       Double min = param.getWarningMin();
-       Double max = param.getWarningMax();
+       Double min = (Double)param.getWarningMin();
+       Double max = (Double)param.getWarningMax();
 
        String name = param.getName();
