+++ org/scec/sha/imr/attenRelImpl/gui/AttenuationRelationshipTesterApp.java
@@ -1282,8 +1282,8 @@ public class AttenuationRelationshipTesterApp extends JApplet
 
 
         try{
-            Double min = param.getWarningMin();
-            Double max = param.getWarningMax();
+            Double min = (Double)param.getWarningMin();
+            Double max = (Double)param.getWarningMax();
 
             String name = param.getName();
